Default debug_file to the code_file basename
Strip any path from code_file, and use that as debug_file if one isn't recorded
in the codeview record.

The current implementation of 'ld --build-id' doesn't write a pdbfilename in the
codeview record, not least because we don't have such a thing.

The minidump processor requires both debug_file and debug_identifier (the
build-id) to locate the symbol file.